Pausing Story 2 of backlog 2 -Getting rid of some magic strings and putting them as an enum
-----------------------------CURRENTLY CUCUMBER TESTS AND UNIT TEST CHECKS---------------------------------
- Fixed add ride cucumber

Next steps:
- PRIORITY: FINISH CHECKING TESTS
- Upload image (deferred)
- Begin Story 4

Notes:
- May be possible in the future to restructure account to hold person, which holds personal details.

+++ src/main/java/controllers/AccountCreationController.java
@@ -105,7 +105,7 @@ public class AccountCreationController extends Controller {
                         nameText.getText(),
                         addressText.getText(),
                         Integer.parseInt(homeNumberText.getText()),
-                        Integer.parseInt(mobileNumberText.getText()),
+                        Long.parseLong(mobileNumberText.getText()),
                         licenceToAdd);
                 content = "You have created a driver account with the User ID: " + universityIDText.getText() +
                         "\nYou will now be taken back to the log in screen.";
@@ -117,7 +117,7 @@ public class AccountCreationController extends Controller {
                         nameText.getText(),
                         addressText.getText(),
                         Integer.parseInt(homeNumberText.getText()),
-                        Integer.parseInt(mobileNumberText.getText()));
+                        Long.parseLong(mobileNumberText.getText()));
                 content = "You have created a passenger account with the User ID: " + universityIDText.getText() +
                         ".\nIf this is incorrect and you had licence details, you can add it in later." +
                         "\nYou will now be taken back to the log in screen.";
